requirements.txt
Django==3.2.7
django-braces==1.14.0
django-cors-headers==3.8.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
docopt==0.6.2
drf-yasg2==1.19.4
filelock==3.0.12
ghp-import==2.0.1
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
mysqlclient==2.0.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
pipenv==2021.5.29
pipreqs==0.4.10
platformdirs==2.3.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyyaml_env_tag==0.1
requests==2.26.0
requests-oauthlib==1.3.0
django-environ==0.7.0
